+++ src/main/java/fr/maxlego08/essentials/commands/CommandLoader.java
@@ -5,6 +5,7 @@ import fr.maxlego08.essentials.api.EssentialsPlugin;
 import fr.maxlego08.essentials.api.commands.CommandManager;
 import dev.yanianz.essentials.nicknames.CommandNick;
 import dev.yanianz.essentials.networkchat.CommandGlobalChat;
+import dev.yanianz.essentials.friends.CommandFriend;
 import dev.yanianz.essentials.reports.CommandReport;
 import dev.yanianz.essentials.reports.CommandReports;
 import dev.yanianz.essentials.notes.CommandNotes;
@@ -131,6 +132,7 @@ public class CommandLoader {
         register("back", CommandTeleportBack.class);
         register("tpr", CommandTeleportRandom.class, "rtp");
         register("g", CommandGlobalChat.class, "globalchat");
+        register("friend", CommandFriend.class, "friends");
 
         register("balancetop", CommandBalanceTop.class, "baltop");
         register("economy", CommandEconomy.class, "eco");